Fully documented with examples.
Added documentation and examples to all scripts.  fun_Helper.R  added
but did not make active as these functions don't need to be exposed.
Put in dummy value in fun.UserDefinedValues.R for what shows up in the
help file.  QC properly referenced "zoo".  Added defaults to many of the
scripts.  Added a full set of examples to the master script.